Adding de method readerToComputer and ExistById in the ComputerRepository class

+++ Program.cs
@@ -3,7 +3,6 @@ using LabManager.Database;
 using LabManager.Repositories;
 using LabManager.Models;
 
-
 var databaseConfig = new DatabaseConfig();
 var databaseSetup = new DatabaseSetup(databaseConfig);
 var computerRepository = new ComputerRepository(databaseConfig);
@@ -35,24 +34,39 @@ if(modelName == "Computer")
     if(modelAction == "Show")
     {
         var id = Convert.ToInt32(args[2]);
+        if(computerRepository.ExistsById(id))
+        {
         var computer = computerRepository.GetById(id);
         Console.WriteLine($"{computer.Id}, {computer.Ram}, {computer.Processor}");
+        } else {
+            Console.WriteLine($"O computador com id {id} não existe");
+        }
     }
 
     if(modelAction == "Update")
     {
         var id = Convert.ToInt32(args[2]);
+        if(computerRepository.ExistsById(id))
+        {
         string ram = args[3];
         string processor = args[4];
 
         var computer = new Computer(id, ram, processor);
         computerRepository.Update(computer);
+        } else {
+            Console.WriteLine($"O computador com id {id} não existe");
+        }
     }
     
     if(modelAction == "Delete")
     {
         var id = Convert.ToInt32(args[2]);
+        if(computerRepository.ExistsById(id))
+        {
         computerRepository.Delete(id);
+        } else {
+            Console.WriteLine($"O computador com id {id} não existe");
+        }
     }
 }
 if(modelName == "Lab")
